requirements.txt
jinja2
pyYAML
junos-eznc
libvirt-python
